+++ app/controllers/projects_controller.rb
@@ -52,7 +52,7 @@ class ProjectsController < ApplicationController
   end
 
   private def project_params
-    params.require(:project).permit(:urn, :trust_ukprn, :delivery_officer_id)
+    params.require(:project).permit(:urn, :trust_ukprn, :target_completion_date, :delivery_officer_id)
   end
 
   private def assign_team_leader