+++ app/models/project.rb
@@ -3,8 +3,10 @@ class Project < ApplicationRecord
 
   validates :urn, presence: true, numericality: {only_integer: true}
   validates :trust_ukprn, presence: true, numericality: {only_integer: true}
+  validates :target_completion_date, presence: true
   validates :team_leader, presence: true
   validate :establishment_exists, :trust_exists, on: :create
+  validate :target_completion_date, :first_day_of_month
 
   belongs_to :delivery_officer, class_name: "User", optional: true
   belongs_to :team_leader, class_name: "User", optional: false
@@ -41,5 +43,14 @@ class Project < ApplicationRecord
     raise result.error if result.error.present?
 
     @trust = result.object
+  end
+
+  private def first_day_of_month
+    return if target_completion_date.nil?
+
+    # Target completion date is always the 1st of the month.
+    if target_completion_date.day != 1
+      errors.add(:target_completion_date, :must_be_first_of_the_month)
+    end
   end
 end