+++ db/migrate/20220804093743_add_target_completion_date_to_project.rb
@@ -1,0 +1,5 @@
+class AddTargetCompletionDateToProject < ActiveRecord::Migration[7.0]
+  def change
+    add_column :projects, :target_completion_date, :date, null: false
+  end
+end