+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Users must provide the UKPRN of the incoming trust when creating a project.
 - The project information view has a Trust details section, which displays the
   incoming trust name, UKPRN, and Companies House number.
+- Users must provide the target completion date when creating a project.
 
 ## [Release 1][release-1]